Player can unequip gear by right clicking it in the inventory menu

+++ Source/Rites/ItemData.h
@@ -21,8 +21,8 @@ public:
 	int32 InstanceID;
 
 	UPROPERTY(EditAnywhere)
-	int32 Count;
+	int16 Count;
 
 	UPROPERTY(EditAnywhere)
-	int32 Durability;
+	int16 Durability;
 };
